requirements.txt
confluent-kafka==2.6.0
kafka-python==2.0.2
